requirements.txt
Django==5.0.2
djangorestframework==3.14.0
requests==2.31.0
beautifulsoup4==4.12.3